requirements.txt
asgiref==3.2.7
Django==3.0.7
django-braces==1.14.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
djangorestframework==3.11.0
markdown2==2.3.9
Pillow==7.1.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
